add baseAmount in Like_List

+++ src/main/java/com/example/finance/entity/Product.java
@@ -3,6 +3,8 @@ package com.example.finance.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 public class Product {
@@ -10,8 +12,7 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long No;    // 產品流水號 (Primary Key)
-
     private String ProductName; //產品名稱
-    private double Price;   //產品價格
-    private double FeeRate; //手續費率(%)
+    private BigDecimal Price;   //產品價格
+    private BigDecimal FeeRate; //手續費率(%)
 }
