+++ src/main/java/com/example/finance/repository/LikeListRepositoryCustom.java
@@ -1,5 +1,5 @@
 package com.example.finance.repository;
 
 public interface LikeListRepositoryCustom {
-    void callAddLikeSP(String userId, Integer productId, Integer quantity, String account);
+    void callAddLikeSP(String userId, Long productId, Integer quantity, String account);
 }
